refactor: registry events to own class
