+++ src/main/java/mrriegel/storagenetwork/StorageNetwork.java
@@ -1,26 +1,13 @@
 package mrriegel.storagenetwork;
 
 import org.apache.logging.log4j.Logger;
-import mrriegel.storagenetwork.block.cable.TileCable;
-import mrriegel.storagenetwork.block.master.TileMaster;
-import mrriegel.storagenetwork.block.request.TileRequest;
 import mrriegel.storagenetwork.config.ConfigHandler;
-import mrriegel.storagenetwork.item.ItemUpgrade;
-import mrriegel.storagenetwork.item.remote.ItemRemote;
 import mrriegel.storagenetwork.proxy.CommonProxy;
-import mrriegel.storagenetwork.registry.ModBlocks;
-import mrriegel.storagenetwork.registry.ModItems;
-import net.minecraft.block.Block;
-import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import mrriegel.storagenetwork.registry.RegistryEvents;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemBlock;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.translation.I18n;
-import net.minecraftforge.client.event.ModelRegistryEvent;
-import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.Mod.Instance;
@@ -28,8 +15,6 @@ import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.registry.GameRegistry;
 
 @Mod(modid = StorageNetwork.MODID, name = StorageNetwork.MODNAME, updateJSON = "https://raw.githubusercontent.com/PrinceOfAmber/Storage-Network/master/update.json")
 public class StorageNetwork {
@@ -42,17 +27,12 @@ public class StorageNetwork {
   @SidedProxy(clientSide = "mrriegel.storagenetwork.proxy.ClientProxy", serverSide = "mrriegel.storagenetwork.proxy.CommonProxy")
   public static CommonProxy proxy;
 
-  public static void log(String s) {
-    if (ConfigHandler.logEverything) {
-      instance.logger.info(s);
-    }
-  }
-
   @EventHandler
   public void preInit(FMLPreInitializationEvent event) {
     logger = event.getModLog();
     proxy.preInit(event);
     MinecraftForge.EVENT_BUS.register(this);
+    MinecraftForge.EVENT_BUS.register(new RegistryEvents());
   }
 
   @EventHandler
@@ -63,47 +43,6 @@ public class StorageNetwork {
   @EventHandler
   public void postInit(FMLPostInitializationEvent event) {
     proxy.postInit(event);
-  }
-
-  @SubscribeEvent
-  public void onRegistryBlock(RegistryEvent.Register<Block> event) {
-    event.getRegistry().register(ModBlocks.master);
-    event.getRegistry().register(ModBlocks.request);
-    event.getRegistry().register(ModBlocks.kabel.setUnlocalizedName(ModBlocks.kabel.getRegistryName().toString()));
-    event.getRegistry().register(ModBlocks.storageKabel.setUnlocalizedName(ModBlocks.storageKabel.getRegistryName().toString()));
-    event.getRegistry().register(ModBlocks.exKabel.setUnlocalizedName(ModBlocks.exKabel.getRegistryName().toString()));
-    event.getRegistry().register(ModBlocks.imKabel.setUnlocalizedName(ModBlocks.imKabel.getRegistryName().toString()));
-  }
-
-  @SubscribeEvent
-  public void onRegistryEvent(RegistryEvent.Register<Item> event) {
-    event.getRegistry().register(new ItemBlock(ModBlocks.master).setRegistryName(ModBlocks.master.getRegistryName()));
-    event.getRegistry().register(new ItemBlock(ModBlocks.request).setRegistryName(ModBlocks.request.getRegistryName()));
-    event.getRegistry().register(new ItemBlock(ModBlocks.kabel).setRegistryName(ModBlocks.kabel.getRegistryName()));
-    event.getRegistry().register(new ItemBlock(ModBlocks.storageKabel).setRegistryName(ModBlocks.storageKabel.getRegistryName()));
-    event.getRegistry().register(new ItemBlock(ModBlocks.exKabel).setRegistryName(ModBlocks.exKabel.getRegistryName()));
-    event.getRegistry().register(new ItemBlock(ModBlocks.imKabel).setRegistryName(ModBlocks.imKabel.getRegistryName()));
-    event.getRegistry().register(ModItems.upgrade);
-    event.getRegistry().register(ModItems.remote.setUnlocalizedName(ModItems.remote.getRegistryName().toString()));
-    GameRegistry.registerTileEntity(TileCable.class, "tileKabel");
-    GameRegistry.registerTileEntity(TileMaster.class, "tileMaster");
-    GameRegistry.registerTileEntity(TileRequest.class, "tileRequest");
-  }
-
-  @SubscribeEvent
-  public void registerModels(ModelRegistryEvent event) {
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.kabel), 0, new ModelResourceLocation(StorageNetwork.MODID + ":kabel", "inventory"));
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.exKabel), 0, new ModelResourceLocation(StorageNetwork.MODID + ":ex_kabel", "inventory"));
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.storageKabel), 0, new ModelResourceLocation(StorageNetwork.MODID + ":storage_kabel", "inventory"));
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.imKabel), 0, new ModelResourceLocation(StorageNetwork.MODID + ":im_kabel", "inventory"));
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.master), 0, new ModelResourceLocation(StorageNetwork.MODID + ":master", "inventory"));
-    ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(ModBlocks.request), 0, new ModelResourceLocation(StorageNetwork.MODID + ":request", "inventory"));
-    for (int i = 0; i < ItemUpgrade.NUM; i++) {
-      ModelLoader.setCustomModelResourceLocation(ModItems.upgrade, i, new ModelResourceLocation(StorageNetwork.MODID + ":upgrade_" + i, "inventory"));
-    }
-    for (int i = 0; i < ItemRemote.RemoteType.values().length; i++) {
-      ModelLoader.setCustomModelResourceLocation(ModItems.remote, i, new ModelResourceLocation(StorageNetwork.MODID + ":remote_" + i, "inventory"));
-    }
   }
 
   public static void chatMessage(EntityPlayer player, String message) {
@@ -121,6 +60,12 @@ public class StorageNetwork {
   }
 
   private static long lastTime;
+
+  public static void log(String s) {
+    if (ConfigHandler.logEverything) {
+      instance.logger.info(s);
+    }
+  }
 
   public static void benchmark(String s) {
     long now = System.currentTimeMillis();
